requirements-test.txt
-r requirements.txt
coverage
coveralls
mock>=1.0.1
model-mommy>=1.2.3
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.7.0

# Additional test requirements go here
